+++ src/Reservation.Domain/Entities/Finances/BusinessRequestPay.cs
@@ -7,7 +7,7 @@ public sealed class BusinessRequestPay : BaseClass<Guid>
     public bool IsPay { get; set; } = false;
     public DateTime PayDate { get; set; }
     public string Authorizy { get; set; } = null!;
-    public int RefId { get; set; }
+    public ulong RefId { get; set; }
 
     // Request Pay Business
     public Business Business { get; set; }
